requirements-minimal.txt
# Core dependencies
flask>=2.0.0
Werkzeug>=2.0.0
python-dotenv>=0.19.0
requests>=2.25.0
numpy>=1.20.0
gunicorn>=20.1.0
flask-cors>=3.0.10

# Machine learning essentials
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
huggingface_hub>=0.10.0

# Audio processing essentials
librosa>=0.8.0
soundfile>=0.10.0
pydub>=0.25.0
demucs>=4.0.0

# API integrations
aiohttp>=3.8.5
asyncio>=3.4.3
aiofiles>=24.1.0
base64io>=1.0.3
google-generativeai>=0.3.0
google-cloud-secret-manager>=2.12.0

# Azure Storage (needed by Sarvam Batch API)
azure-storage-file-datalake>=12.18.0
azure-identity>=1.12.0
azure-core>=1.26.0

# Validation and metrics
jiwer>=2.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
pandas>=1.3.0
sentence-transformers>=2.2.2
sacrebleu>=2.3.1

# Video processing
yt-dlp>=2023.3.4
pytube>=15.0.0
google-api-python-client>=2.100.0
youtube-dl-server>=0.3
